I think I'm done

+++ routes/api/user/user.js
@@ -22,7 +22,7 @@ router.post("/users", async (req, res) => {
         await registerUserValidation(newUser);
         newUser.password = await hashService.generateHash(newUser.password);
         await userServiceModel.registerUser(newUser);
-        res.status(201).json("user registered");
+        res.status(201).json({ userRegistered: newUser });
     }
     catch (err) {
         res.status(400).json({ message: err.message || err });
@@ -37,7 +37,6 @@ router.post("/users/login", async (req, res) => {
             throw new Error("Invalid user name or password. please try again!");
         }
         const validatePassword = await hashService.compareHash(req.body.password, currentUser.password)
-        console.log("validate password", validatePassword);
         if (!validatePassword) {
             throw new Error("Invalid user name or password. please try again!")
         }
@@ -90,11 +89,16 @@ router.get("/:id", loggedInCheck, async (req, res) => {
 
 router.put("/:id", loggedInCheck, async (req, res) => {
     try {
-        await validateId({ id: req.params.id });
-        await validateEditUser(req.body);
         const userId = req.params.id
+        await validateId({ id: req.params.id });
+        const userToEdit = await userServiceModel.getUserById(userId)
+        if (!userToEdit) {
+            throw ("user does not exist");
+        }
+        await validateEditUser(req.body);
         if (userId === req.tokenPayload.userId.toString()) {
             delete req.body.isBusiness;
+            delete req.body.password;
             await userServiceModel.updateUser(req.params.id, req.body);
             return res.status(200).json({ message: "Update Successful" });
         }
@@ -112,6 +116,9 @@ router.patch("/:id", loggedInCheck, async (req, res) => {
         await validateId({ id: req.params.id });
         const userId = req.params.id;
         const userProfile = await userServiceModel.getUserById(userId);
+        if (!userProfile) {
+            throw ("User does not exist")
+        }
         if (userId === req.tokenPayload.userId.toString()) {
             userProfile.isBusiness = !userProfile.isBusiness;
             await userServiceModel.updateUser(userId, userProfile);
@@ -120,7 +127,6 @@ router.patch("/:id", loggedInCheck, async (req, res) => {
         else throw { message: "You are not the user you are trying to edit!" }
     }
     catch (err) {
-        console.log("err from isBusiness Status Change", err.message || err);
         res.status(401).json({ message: err.message || err });
     }
 })
@@ -130,6 +136,10 @@ router.patch("/:id", loggedInCheck, async (req, res) => {
 router.delete("/:id", loggedInCheck, async (req, res) => {
     try {
         await validateId({ id: req.params.id });
+        const idToDelete = await userServiceModel.getUserById(req.params.id);
+        if (!idToDelete) {
+            throw ("user does not exist")
+        }
         if (req.tokenPayload.isAdmin || req.tokenPayload.userId.toString() === req.params.id) {
             await userServiceModel.deleteUser(req.params.id)
             res.status(200).json({ message: "user deleted!" })
